LayoutTests/imported/w3c/web-platform-tests/trusted-types/block-Node-multiple-arguments-expected.txt

CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
'createScript';
'createScript';
'createScript';
'createScript';
'createScript';
'createScript #1';'#2;'
'createScript #1';'#2;'
'createScript #1';'#2;'
'createScript #1';'#2;'
'createScript #1';'#2;'
'plain text';
'plain text';
'plain text';
'plain text';
'plain text';
'plain text #1';'plain text #2';
'plain text #1';'plain text #2';
'plain text #1';'plain text #2';
'plain text #1';'plain text #2';
'plain text #1';'plain text #2';
'node';
'node #1';'node #2';
'mixed';
'mixed';
'mixed';
'mixed';
'mixed';'node';
'mixed';'script';
'mixed';'script';
'mixed';'script';
'mixed';'script';
'mixed';'script';
'script';
'script';
'script';
'script';
'node';'script';

PASS replaceWith('createScript';) on <div> should pass
PASS after('createScript';) on <div> should pass
PASS before('createScript';) on <div> should pass
PASS append('createScript';) on <div> should pass
PASS prepend('createScript';) on <div> should pass
PASS replaceWith('createScript #1';,'#2;') on <div> should pass
PASS after('createScript #1';,'#2;') on <div> should pass
PASS before('createScript #1';,'#2;') on <div> should pass
PASS append('createScript #1';,'#2;') on <div> should pass
PASS prepend('createScript #1';,'#2;') on <div> should pass
PASS replaceWith('plain text';) on <div> should pass
PASS after('plain text';) on <div> should pass
PASS before('plain text';) on <div> should pass
PASS append('plain text';) on <div> should pass
PASS prepend('plain text';) on <div> should pass
PASS replaceWith('plain text #1';,'plain text #2';) on <div> should pass
PASS after('plain text #1';,'plain text #2';) on <div> should pass
PASS before('plain text #1';,'plain text #2';) on <div> should pass
PASS append('plain text #1';,'plain text #2';) on <div> should pass
PASS prepend('plain text #1';,'plain text #2';) on <div> should pass
PASS replaceWith([object Text]) on <div> should pass
PASS after([object Text]) on <div> should pass
PASS before([object Text]) on <div> should pass
PASS append([object Text]) on <div> should pass
PASS prepend([object Text]) on <div> should pass
PASS replaceWith([object Text],[object Text]) on <div> should pass
PASS after([object Text],[object Text]) on <div> should pass
PASS before([object Text],[object Text]) on <div> should pass
PASS append([object Text],[object Text]) on <div> should pass
PASS prepend([object Text],[object Text]) on <div> should pass
PASS replaceWith('mixed';,[object Text]) on <div> should pass
PASS after('mixed';,[object Text]) on <div> should pass
PASS before('mixed';,[object Text]) on <div> should pass
PASS append('mixed';,[object Text]) on <div> should pass
PASS prepend('mixed';,[object Text]) on <div> should pass
PASS replaceWith('mixed';,'script';) on <div> should pass
PASS after('mixed';,'script';) on <div> should pass
PASS before('mixed';,'script';) on <div> should pass
PASS append('mixed';,'script';) on <div> should pass
PASS prepend('mixed';,'script';) on <div> should pass
PASS replaceWith([object Text],'script';) on <div> should pass
PASS after([object Text],'script';) on <div> should pass
PASS before([object Text],'script';) on <div> should pass
PASS append([object Text],'script';) on <div> should pass
PASS prepend([object Text],'script';) on <div> should pass
PASS replaceWith('createScript';) on <script> should pass
PASS after('createScript';) on <script> should pass
PASS before('createScript';) on <script> should pass
PASS append('createScript';) on <script> should pass
PASS prepend('createScript';) on <script> should pass
PASS replaceWith('createScript #1';,'#2;') on <script> should pass
PASS after('createScript #1';,'#2;') on <script> should pass
PASS before('createScript #1';,'#2;') on <script> should pass
PASS append('createScript #1';,'#2;') on <script> should pass
PASS prepend('createScript #1';,'#2;') on <script> should pass
PASS replaceWith('plain text';) on <script> should fail
PASS after('plain text';) on <script> should fail
PASS before('plain text';) on <script> should fail
PASS append('plain text';) on <script> should fail
PASS prepend('plain text';) on <script> should fail
PASS replaceWith('plain text #1';,'plain text #2';) on <script> should fail
PASS after('plain text #1';,'plain text #2';) on <script> should fail
PASS before('plain text #1';,'plain text #2';) on <script> should fail
PASS append('plain text #1';,'plain text #2';) on <script> should fail
PASS prepend('plain text #1';,'plain text #2';) on <script> should fail
PASS replaceWith([object Text]) on <script> should fail
PASS after([object Text]) on <script> should fail
PASS before([object Text]) on <script> should fail
PASS append([object Text]) on <script> should fail
PASS prepend([object Text]) on <script> should fail
PASS replaceWith([object Text],[object Text]) on <script> should fail
PASS after([object Text],[object Text]) on <script> should fail
PASS before([object Text],[object Text]) on <script> should fail
PASS append([object Text],[object Text]) on <script> should fail
PASS prepend([object Text],[object Text]) on <script> should fail
PASS replaceWith('mixed';,[object Text]) on <script> should fail
PASS after('mixed';,[object Text]) on <script> should fail
PASS before('mixed';,[object Text]) on <script> should fail
PASS append('mixed';,[object Text]) on <script> should fail
PASS prepend('mixed';,[object Text]) on <script> should fail
PASS replaceWith('mixed';,'script';) on <script> should fail
PASS after('mixed';,'script';) on <script> should fail
PASS before('mixed';,'script';) on <script> should fail
PASS append('mixed';,'script';) on <script> should fail
PASS prepend('mixed';,'script';) on <script> should fail
PASS replaceWith([object Text],'script';) on <script> should fail
PASS after([object Text],'script';) on <script> should fail
PASS before([object Text],'script';) on <script> should fail
PASS append([object Text],'script';) on <script> should fail
PASS prepend([object Text],'script';) on <script> should fail
